Pull coin out into own module
