+++ packages/std/src/coins.rs
@@ -1,0 +1,16 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
+pub struct Coin {
+    pub denom: String,
+    pub amount: String,
+}
+
+// coin is a shortcut constructor for a set of one denomination of coins
+pub fn coin(amount: &str, denom: &str) -> Vec<Coin> {
+    vec![Coin {
+        amount: amount.to_string(),
+        denom: denom.to_string(),
+    }]
+}
